WinPython registration: now registers entries in HKCU\Software\Python\PythonCore
In other words, any distutils Python package will see WinPython distribution as
a standard installed Python distribution.

+++ winpython/associate.py
@@ -23,12 +23,17 @@ from winpython.py3compat import winreg
 from winpython import utils
 
 KEY_C = r"Software\Classes\%s"
-KEY_C0 = KEY_C % r"Python.%sFile\shell\%s"
-KEY_C1 = KEY_C0 + r"\command"
+KEY_C0 = KEY_C % r"Python.%sFile\shell"
+KEY_C1 = KEY_C % r"Python.%sFile\shell\%s"
+KEY_C2 = KEY_C1 + r"\command"
 KEY_I = KEY_C % r"Python.%sFile\DefaultIcon"
 KEY_D = KEY_C % r"Python.%sFile"
 EWI = "Edit with IDLE"
 EWS = "Edit with Spyder"
+
+KEY_S = r"Software\Python"
+KEY_S0 = KEY_S + r"\PythonCore"
+KEY_S1 = KEY_S0 + r"\%s"
 
 
 def _get_shortcut_data(target, current=True):
@@ -67,24 +72,24 @@ def register(target, current=True):
     # Verbs
     python = osp.abspath(osp.join(target, 'python.exe'))
     pythonw = osp.abspath(osp.join(target, 'pythonw.exe'))
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("", "open")),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("", "open")),
                       "", 0, winreg.REG_SZ, '"%s" "%%1" %%*' % python)
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("NoCon", "open")),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("NoCon", "open")),
                       "", 0, winreg.REG_SZ, '"%s" "%%1" %%*' % pythonw)
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("Compiled", "open")),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("Compiled", "open")),
                       "", 0, winreg.REG_SZ, '"%s" "%%1" %%*' % python)
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("", EWI)),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("", EWI)),
                       "", 0, winreg.REG_SZ,
                       '"%s" "%s\Lib\idlelib\idle.pyw" -n -e "%%1"'
                       % (pythonw, target))
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("NoCon", EWI)),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("NoCon", EWI)),
                       "", 0, winreg.REG_SZ,
                       '"%s" "%s\Lib\idlelib\idle.pyw" -n -e "%%1"'
                       % (pythonw, target))
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("", EWS)),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("", EWS)),
                       "", 0, winreg.REG_SZ,
                       '"%s" "%s\Scripts\spyder" "%%1"' % (pythonw, target))
-    winreg.SetValueEx(winreg.CreateKey(root, KEY_C1 % ("NoCon", EWS)),
+    winreg.SetValueEx(winreg.CreateKey(root, KEY_C2 % ("NoCon", EWS)),
                       "", 0, winreg.REG_SZ,
                       '"%s" "%s\Scripts\spyder" "%%1"' % (pythonw, target))
     
@@ -105,6 +110,25 @@ def register(target, current=True):
     winreg.SetValueEx(winreg.CreateKey(root, KEY_D % "Compiled"),
                       "", 0, winreg.REG_SZ, "Compiled Python File")
 
+    # PythonCore entries
+    short_version = utils.get_python_infos(target)[0]
+    long_version = utils.get_python_long_version(target)
+    key_core = (KEY_S1 % short_version) + r'\%s'
+    winreg.SetValueEx(winreg.CreateKey(root, key_core % 'InstallPath'),
+                      "", 0, winreg.REG_SZ, target)
+    winreg.SetValueEx(winreg.CreateKey(root,
+                                       key_core % r'InstallPath\InstallGroup'),
+                      "", 0, winreg.REG_SZ, "Python %s" % short_version)
+    winreg.SetValueEx(winreg.CreateKey(root, key_core % 'Modules'),
+                      "", 0, winreg.REG_SZ, "")
+    winreg.SetValueEx(winreg.CreateKey(root, key_core % 'PythonPath'),
+                      "", 0, winreg.REG_SZ,
+                      r"%s\Lib;%s\DLLs" % (target, target))
+    winreg.SetValueEx(winreg.CreateKey(root,
+                               key_core % r'Help\Main Python Documentation'),
+                      "", 0, winreg.REG_SZ,
+                      r"%s\Doc\python%s.chm" % (target, long_version))
+
     # Create start menu entries for all WinPython launchers
     for path, desc, fname in _get_shortcut_data(target, current=current):
         utils.create_shortcut(path, desc, fname)
@@ -123,24 +147,38 @@ def unregister(target, current=True):
     """Unregister a Python distribution in Windows registry"""
     # Registry entries
     root = winreg.HKEY_CURRENT_USER if current else winreg.HKEY_LOCAL_MACHINE
+    short_version = utils.get_python_infos(target)[0]
+    key_core = (KEY_S1 % short_version) + r'\%s'
     for key in (
                 # Icons
                 KEY_I % "NoCon", KEY_I % "Compiled", KEY_I % "",
                 # Edit with IDLE
+                KEY_C2 % ("", EWI), KEY_C2 % ("NoCon", EWI),
                 KEY_C1 % ("", EWI), KEY_C1 % ("NoCon", EWI),
-                KEY_C0 % ("", EWI), KEY_C0 % ("NoCon", EWI),
                 # Edit with Spyder
+                KEY_C2 % ("", EWS), KEY_C2 % ("NoCon", EWS),
                 KEY_C1 % ("", EWS), KEY_C1 % ("NoCon", EWS),
-                KEY_C0 % ("", EWS), KEY_C0 % ("NoCon", EWS),
                 # Verbs
-                KEY_C1 % ("", "open"), KEY_C1 % ("NoCon", "open"),
+                KEY_C2 % ("", "open"),
+                KEY_C2 % ("NoCon", "open"),
+                KEY_C2 % ("Compiled", "open"),
+                KEY_C1 % ("", "open"),
+                KEY_C1 % ("NoCon", "open"),
                 KEY_C1 % ("Compiled", "open"),
+                KEY_C0 % "", KEY_C0 % "NoCon", KEY_C0 % "Compiled",
                 # Descriptions
                 KEY_D % "NoCon", KEY_D % "Compiled", KEY_D % "",
-                # Parent key
-                KEY_C,
+                # PythonCore
+                key_core % r'InstallPath\InstallGroup',
+                key_core % 'InstallPath',
+                key_core % 'Modules',
+                key_core % 'PythonPath',
+                key_core % r'Help\Main Python Documentation',
+                key_core % 'Help',
+                KEY_S1 % short_version, KEY_S0, KEY_S,
                 ):
         try:
+            print(key)
             winreg.DeleteKey(root, key)
         except WindowsError:
             rootkey = 'HKEY_CURRENT_USER' if current else 'HKEY_LOCAL_MACHINE'
@@ -148,7 +186,8 @@ def unregister(target, current=True):
     
     # Start menu shortcuts
     for path, desc, fname in _get_shortcut_data(target, current=current):
-        os.remove(fname)
+        if osp.exists(fname):
+            os.remove(fname)
 
 
 if __name__ == '__main__':
